blog/requirements.txt
# Para trabajar con AWS S3
boto3==1.24.92
botocore==1.27.92

# ASGI
channels==3.0.5
channels-redis==3.4.1

# Para trabajar con el event bus kafka
confluent-kafka==1.9.2

Django==3.2.16
# Esta libreria es para mejorar los response, creada por la persona del tutorial(SoloPython)
djangorestframework-api-response==0.0.33

# Para crear la api
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# cors
django-cors-headers==3.13.0

# Para las variables de ambiente
django-environ==0.9.0

# Para la comunicacion con redis
django-redis==5.2.0

# Para poder guardar videos e imagenes
django-storages==1.13.1

# Para procesar imagenes
Pillow==9.2.0

# Para la comunicacion con postgresql
psycopg2==2.9.4
psycopg-binary==3.1.12

# Para hacer slug
python-slugify==6.1.2

PyJWT==2.5.0

whitenoise==6.2.0

uvicorn==0.20.0

rsa==4.9

jsonpickle==3.0.1
